refactor: downgrade method , update snapshots, and removed useless parameter
