+++ tests/snapshots/nanos_downgrade_ETHx/00004.png
