+++ tests/snapshots/nanos_downgrade_ETHx/00007.png
